requirements.txt
matplotlib==3.7.4
pandas==2.0.3
opensmile==2.5.0
python-dotenv~=1.0.1